services/auth-service/requirements.txt
# Dependências do auth
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
fastapi-mail==1.4.1
